dmaengine: added MIC X200 Coprocessor DMA driver
This patch implements the DMA Engine driver for the DMA controller on
MIC X200 Coprocessors which are PCIe cards running Linux. Separate but
identical DMA channels are available for the host and card. DMA
transfers can be done from both the host and card in host<->card and
host<->host directions.

Signed-off-by: Jacek Lawrynowicz <jacek.lawrynowicz@intel.com>
Signed-off-by: Sudeep Dutt <sudeep.dutt@intel.com>
Signed-off-by: Ashutosh Dixit <ashutosh.dixit@intel.com>
Signed-off-by: Andrzej Kacprowski <andrzej.kacprowski@intel.com>